+++ src/providers/twitch/TwitchIrcServer.hpp
@@ -7,7 +7,6 @@
 #include "providers/bttv/BttvEmotes.hpp"
 #include "providers/ffz/FfzEmotes.hpp"
 #include "providers/irc/AbstractIrcServer.hpp"
-#include "providers/twitch/TwitchBadges.hpp"
 
 #include <chrono>
 #include <memory>
@@ -75,7 +74,6 @@ private:
     std::chrono::steady_clock::time_point lastErrorTimeSpeed_;
     std::chrono::steady_clock::time_point lastErrorTimeAmount_;
 
-    TwitchBadges twitchBadges;
     BttvEmotes bttv;
     FfzEmotes ffz;