+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unversioned
 
+- Major: Added the ability to highlight messages based on user badges. (#1704)
 - Minor: Added visual indicator to message length if over 500 characters long (#2659)
 - Minor: Added `is:<flags>` search filter to find messages of specific types. (#2653, #2671)
 - Minor: Added image links to the badge context menu. (#2667)